Skip target existence checks
so that we can have a podfile which generates xcconfigs
for targets which don't necessarily exist in the host project.

This does have a considerable impact, as some of the functionality
in cocoapods can be affected. Have a look at the diff in this commit
to understand the impacted areas